// src/api/updateAccountSettingsRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tubester API
 * To access protected endpoints, first log in:

[🔐 read only Login with Google](/api/auth/login/google?returnUrl=/swagger/index.html)

[🔐 write Login with Google](/api/auth/login/google/write?returnUrl=/swagger/index.html)
 * OpenAPI spec version: v1
 */

export interface UpdateAccountSettingsRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  preferredTheme: string;
  /**
   * @minLength 1
   * @maxLength 20
   */
  preferredLanguage: string;
}